+++ AutoBid_BE/src/main/java/com/codesquad/autobid/config/kafka/KafkaProducerConfig.java
@@ -1,0 +1,61 @@
+package com.codesquad.autobid.config.kafka;
+
+import lombok.extern.slf4j.Slf4j;
+import org.apache.kafka.common.serialization.StringSerializer;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.annotation.EnableKafka;
+import org.springframework.kafka.core.DefaultKafkaProducerFactory;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.core.ProducerFactory;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.apache.kafka.clients.producer.ProducerConfig.*;
+
+@Configuration
+@EnableKafka
+@Slf4j
+public class ProducerConfig {
+
+    @Value("${spring.kafka.bootstrap-servers}")
+    private String BOOTSTRAP_SERVERS;
+
+    private Map<String, Object> senderProps() {
+        Map<String, Object> props = new HashMap<>();
+        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
+        props.put(ProducerConfig.LINGER_MS_CONFIG, 10);
+        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, IntegerSerializer.class);
+        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+        //...
+        return props;
+    }
+
+    @Bean
+    public ProducerFactory<Integer, String> producerFactory() {
+        return new DefaultKafkaProducerFactory<>(senderProps());
+    }
+
+    @Bean
+    public KafkaTemplate<String, String> kafkaTemplate() {
+        log.info("gigihihihi");
+        Map<String, Object> props = new HashMap<>();
+        props.put(BOOTSTRAP_SERVERS_CONFIG, BOOTSTRAP_SERVERS);
+        props.put(KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+        props.put(VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+
+        return new KafkaTemplate<>(new DefaultKafkaProducerFactory<>(props));
+    }
+
+    // @Bean
+    // public KafkaTemplate<String, Auction> auctionKafkaTemplate() {
+    //     Map<String, Object> props = new HashMap<>();
+    //     props.put(BOOTSTRAP_SERVERS_CONFIG, BOOTSTRAP_SERVERS);
+    //     props.put(KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+    //     props.put(VALUE_SERIALIZER_CLASS_CONFIG, AuctionKafkaDTOSerializer.class);
+//
+    //     return new KafkaTemplate<>(new DefaultKafkaProducerFactory<>(props));
+    // }
+}
