--- AutoBid_BE/src/main/java/com/codesquad/autobid/config/kafka/ProducerConfig.java
@@ -1,42 +1,0 @@
-package com.codesquad.autobid.config.kafka;
-
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static org.apache.kafka.clients.producer.ProducerConfig.*;
-
-@Configuration
-@EnableKafka
-public class ProducerConfig {
-
-    @Value("${spring.kafka.bootstrap-servers}")
-    private String BOOTSTRAP_SERVERS;
-
-    @Bean
-    public ProducerFactory<String, String> producerFactory() {
-        return new DefaultKafkaProducerFactory<>(config());
-    }
-
-    @Bean
-    public Map<String, Object> config() {
-        Map<String, Object> props = new HashMap<>();
-        props.put(BOOTSTRAP_SERVERS_CONFIG, BOOTSTRAP_SERVERS);
-        props.put(KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        props.put(VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        return props;
-    }
-
-    @Bean
-    public KafkaTemplate<String, String> kafkaTemplate() {
-        return new KafkaTemplate<>(producerFactory());
-    }
-}
